--- a/src/Hashids.net/Alphabets/IAlphabetProvider.cs
+++ b/src/Hashids.net/Alphabets/IAlphabetProvider.cs
@@ -1,0 +1,9 @@
+﻿namespace HashidsNet.Alphabets
+{
+    public interface IAlphabetProvider
+    {
+        IAlphabet GetAlphabet(int index);
+        IAlphabet GetAlphabet(char @char);
+        IAlphabet Default { get; }
+    }
+}