--- a/src/Hashids.net/Alphabets/Lottery/LotteryAlphabet.PageSalt.cs
+++ b/src/Hashids.net/Alphabets/Lottery/LotteryAlphabet.PageSalt.cs
@@ -1,7 +1,7 @@
 ﻿using HashidsNet.Alphabets.Salts;
 using System;
 
-namespace HashidsNet.Alphabets
+namespace HashidsNet.Alphabets.Lottery
 {
     public partial class LotteryAlphabet
     {