--- a/src/Hashids.net/Alphabets/Lottery/LotteryAlphabet.ShuffleSalt.cs
+++ b/src/Hashids.net/Alphabets/Lottery/LotteryAlphabet.ShuffleSalt.cs
@@ -1,6 +1,6 @@
 ﻿using HashidsNet.Alphabets.Salts;
 
-namespace HashidsNet.Alphabets
+namespace HashidsNet.Alphabets.Lottery
 {
     public partial class LotteryAlphabet
     {