--- a/src/Hashids.net/Alphabets/Salts/ISalt.cs
+++ b/src/Hashids.net/Alphabets/Salts/ISalt.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace HashidsNet.Alphabets.Salts
+{
+    public interface ISalt
+    {
+        void Calculate(Span<int> buffer, int saltIndex, ref int saltSum);
+
+        int Length { get; }
+    }
+}