--- a/src/Hashids.net/Alphabets/Salts/ISalt.cs
+++ b/src/Hashids.net/Alphabets/Salts/ISalt.cs
@@ -4,7 +4,7 @@ namespace HashidsNet.Alphabets.Salts
 {
     public interface ISalt
     {
-        void Calculate(Span<int> buffer, int saltIndex, ref int saltSum);
+        void Calculate(Span<int> buffer, ref int saltSum);
 
         int Length { get; }
     }