… src/Hashids.net/Alphabets/Salts/WeekSalt.cs
… src/Hashids.net/Alphabets/StepsAlphabet.cs
… src/Hashids.net/FastMods/FastMod.cs
… src/Hashids.net/FastMods/Mod.cs
… src/Hashids.net/FastMods/Mods.cs
… src/Hashids.net/Hashids.EncodingContext.cs
… src/Hashids.net/Hashids.HashDecoder.cs
… src/Hashids.net/Hashids.HashEncoder.cs
… src/Hashids.net/Hashids.HashStats.cs
… src/Hashids.net/Hashids.IdleWriter.cs
… src/Hashids.net/Hashids.PayloadWriter.cs
… src/Hashids.net/Hashids.cs
… src/Hashids.net/Prev/Hashids.Obsolete.cs
… src/Hashids.net/Prev/Hashids.cs
… test/Hashids.net.benchmark/ABBenchmarks.cs
… test/Hashids.net.benchmark/DecodeBenchmarks.cs
… test/Hashids.net.benchmark/DecodeLongBenchmarks.cs
… test/Hashids.net.benchmark/DecodeSingleBenchmarks.cs
… test/Hashids.net.benchmark/DecodeSingleLongBenchmarks.cs
… test/Hashids.net.benchmark/EncodeBenchmarks.cs
… test/Hashids.net.benchmark/EncodeLongBenchmarks.cs
… test/Hashids.net.benchmark/EncodeLongParamsBenchmarks.cs
… test/Hashids.net.benchmark/EncodeParamsBenchmarks.cs
… test/Hashids.net.benchmark/HashBenchmarks.cs
… test/Hashids.net.benchmark/HashidsVersion.cs
… test/Hashids.net.benchmark/Program.cs