I would like to ask merge the changes that improves performance & memory allocation.

## Reason

The original solution is fast enough for any purposes.
But, while I investigate the project I have found some points that could be improved:
* Reduce memory allocation for all encode cases;
* Zero-allocation for DecodeSingle;
* Don't encode numbers while decoding them;
* Cache a few first Alphabets;
* Split the complicated logic on few blocks.

## About the structure

I introduced a few key objects `ISalt`,  `IAlphabet`, `HashEncoder` and `HashDecoder`.

The `ISalt` object is responsible to hold data about salt and calculates it when requested.
A `SaltHelper` provides static method to work with the `ISalt` object. 
You can snapshot a salt object to cache calculation results and/or you can shuffle the existing array.

The `IAlphabet` objects provides interface to interact with alphabet.
The object was introduced to make different flows possible.
For ex.: we can use `LotteryAlphabet` alphabet to head-allocation-safe-scenarios, or use `StepsAlphabet` where the next page of shuffle is cached.
Using a combination of these objects we can implement cacheable or non-cacheable version of `Hashids`.

The `HashEncoder` and `HashDecoder` and its dependencies objects were created to convert the `GenerateHashFrom` and `GetNumbersFrom` to the [Method Objects](https://refactoring.guru/replace-method-with-method-object).
All these objects doesn't allocate a space in heap and use only stack.
The object like `HashStats` and `IdleWriter` shared for encode and decode scenarios. 
There are some hucks inside these objects, they all dictated due to struct behavior.

## Benchmarks result

I saved the previous implementation in the `Prev` folder.
The `HashBenchmarks` was modified to be able compare versions.
Here is a result:

``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.22000
Intel Core i5-10210U CPU 1.60GHz, 1 CPU, 8 logical and 4 physical cores
.NET SDK=6.0.201
  [Host]     : .NET 6.0.3 (6.0.322.12309), X64 RyuJIT
  Job-UBOQKK : .NET 5.0.15 (5.0.1522.11506), X64 RyuJIT
  Job-XQYQUJ : .NET 6.0.3 (6.0.322.12309), X64 RyuJIT
  Job-JLANOM : .NET Framework 4.8 (4.8.4470.0), X64 RyuJIT

OutlierMode=DontRemove  MemoryRandomization=True  

```
|               Method |        Job |            Runtime |      Version |       Mean |     Error |    StdDev |     Median | Code Size |  Gen 0 | Allocated |
|--------------------- |----------- |------------------- |------------- |-----------:|----------:|----------:|-----------:|----------:|-------:|----------:|
|        **RoundtripInts** | **Job-UBOQKK** |           **.NET 5.0** |         **Prev** | **4,522.0 ns** |  **90.01 ns** | **263.97 ns** | **4,457.5 ns** |      **62 B** | **0.1602** |     **512 B** |
|        **RoundtripInts** | **Job-UBOQKK** |           **.NET 5.0** |      **Current** | **1,191.2 ns** |  **34.30 ns** | **101.15 ns** | **1,128.4 ns** |      **62 B** | **0.0629** |     **200 B** |
|        **RoundtripInts** | **Job-UBOQKK** |           **.NET 5.0** | **WithoutCache** | **2,854.4 ns** |  **34.28 ns** |  **32.06 ns** | **2,838.6 ns** |      **62 B** | **0.0610** |     **200 B** |
|                      |            |                    |              |            |           |           |            |           |        |           |
|       **RoundtripLongs** | **Job-UBOQKK** |           **.NET 5.0** |         **Prev** | **5,040.2 ns** |  **62.97 ns** |  **58.90 ns** | **5,057.5 ns** |      **62 B** | **0.1602** |     **512 B** |
|       **RoundtripLongs** | **Job-UBOQKK** |           **.NET 5.0** |      **Current** | **1,591.2 ns** |  **18.87 ns** |  **17.65 ns** | **1,590.1 ns** |      **62 B** | **0.0420** |     **136 B** |
|       **RoundtripLongs** | **Job-UBOQKK** |           **.NET 5.0** | **WithoutCache** | **3,248.9 ns** |  **45.52 ns** |  **42.58 ns** | **3,234.0 ns** |      **62 B** | **0.0420** |     **136 B** |
|                      |            |                    |              |            |           |           |            |           |        |           |
|         **RoundtripHex** | **Job-UBOQKK** |           **.NET 5.0** |         **Prev** | **4,257.6 ns** |  **81.81 ns** |  **76.52 ns** | **4,239.8 ns** |      **62 B** | **0.4272** |   **1,344 B** |
|         **RoundtripHex** | **Job-UBOQKK** |           **.NET 5.0** |      **Current** | **2,069.4 ns** |  **33.96 ns** |  **31.76 ns** | **2,063.8 ns** |      **62 B** | **0.3357** |   **1,064 B** |
|         **RoundtripHex** | **Job-UBOQKK** |           **.NET 5.0** | **WithoutCache** | **3,256.7 ns** |  **21.13 ns** |  **19.77 ns** | **3,251.5 ns** |      **62 B** | **0.3357** |   **1,064 B** |
|                      |            |                    |              |            |           |           |            |           |        |           |
|         **SingleNumber** | **Job-UBOQKK** |           **.NET 5.0** |         **Prev** | **1,006.0 ns** |  **14.60 ns** |  **13.66 ns** | **1,008.4 ns** |      **65 B** | **0.0191** |      **64 B** |
|         **SingleNumber** | **Job-UBOQKK** |           **.NET 5.0** |      **Current** |   **181.8 ns** |   **2.52 ns** |   **2.36 ns** |   **181.4 ns** |      **65 B** | **0.0203** |      **64 B** |
|         **SingleNumber** | **Job-UBOQKK** |           **.NET 5.0** | **WithoutCache** |   **804.3 ns** |   **4.75 ns** |   **4.44 ns** |   **803.3 ns** |      **65 B** | **0.0200** |      **64 B** |
|                      |            |                    |              |            |           |           |            |           |        |           |
| **SingleNumberAsParams** | **Job-UBOQKK** |           **.NET 5.0** |         **Prev** | **1,026.1 ns** |  **20.17 ns** |  **26.23 ns** | **1,025.2 ns** |     **124 B** | **0.0496** |     **160 B** |
| **SingleNumberAsParams** | **Job-UBOQKK** |           **.NET 5.0** |      **Current** |   **229.0 ns** |   **1.76 ns** |   **1.65 ns** |   **228.8 ns** |     **124 B** | **0.0508** |     **160 B** |
| **SingleNumberAsParams** | **Job-UBOQKK** |           **.NET 5.0** | **WithoutCache** |   **756.2 ns** |  **15.14 ns** |  **28.43 ns** |   **766.5 ns** |     **124 B** | **0.0505** |     **160 B** |
|                      |            |                    |              |            |           |           |            |           |        |           |
|        **RoundtripInts** | **Job-XQYQUJ** |           **.NET 6.0** |         **Prev** | **4,324.1 ns** |  **48.60 ns** |  **45.46 ns** | **4,298.3 ns** |      **62 B** | **0.1602** |     **512 B** |
|        **RoundtripInts** | **Job-XQYQUJ** |           **.NET 6.0** |      **Current** | **1,083.1 ns** |   **7.28 ns** |   **6.81 ns** | **1,085.3 ns** |      **62 B** | **0.0629** |     **200 B** |
|        **RoundtripInts** | **Job-XQYQUJ** |           **.NET 6.0** | **WithoutCache** | **2,367.0 ns** |  **46.33 ns** |  **72.13 ns** | **2,402.8 ns** |      **62 B** | **0.0610** |     **200 B** |
|                      |            |                    |              |            |           |           |            |           |        |           |
|       **RoundtripLongs** | **Job-XQYQUJ** |           **.NET 6.0** |         **Prev** | **4,974.9 ns** |  **72.70 ns** |  **68.00 ns** | **4,991.6 ns** |      **62 B** | **0.1602** |     **512 B** |
|       **RoundtripLongs** | **Job-XQYQUJ** |           **.NET 6.0** |      **Current** | **1,605.4 ns** |  **24.69 ns** |  **23.09 ns** | **1,601.7 ns** |      **62 B** | **0.0420** |     **136 B** |
|       **RoundtripLongs** | **Job-XQYQUJ** |           **.NET 6.0** | **WithoutCache** | **2,756.7 ns** |  **54.58 ns** |  **81.70 ns** | **2,805.3 ns** |      **62 B** | **0.0420** |     **136 B** |
|                      |            |                    |              |            |           |           |            |           |        |           |
|         **RoundtripHex** | **Job-XQYQUJ** |           **.NET 6.0** |         **Prev** | **4,123.1 ns** |  **34.88 ns** |  **32.63 ns** | **4,119.0 ns** |      **62 B** | **0.4272** |   **1,344 B** |
|         **RoundtripHex** | **Job-XQYQUJ** |           **.NET 6.0** |      **Current** | **2,024.6 ns** |  **28.47 ns** |  **26.63 ns** | **2,017.7 ns** |      **62 B** | **0.3357** |   **1,064 B** |
|         **RoundtripHex** | **Job-XQYQUJ** |           **.NET 6.0** | **WithoutCache** | **2,681.0 ns** |  **52.14 ns** |  **78.05 ns** | **2,645.5 ns** |      **62 B** | **0.3357** |   **1,064 B** |
|                      |            |                    |              |            |           |           |            |           |        |           |
|         **SingleNumber** | **Job-XQYQUJ** |           **.NET 6.0** |         **Prev** |   **987.6 ns** |  **19.45 ns** |  **19.10 ns** |   **988.9 ns** |      **65 B** | **0.0191** |      **64 B** |
|         **SingleNumber** | **Job-XQYQUJ** |           **.NET 6.0** |      **Current** |   **174.9 ns** |   **3.22 ns** |   **3.02 ns** |   **175.0 ns** |      **65 B** | **0.0203** |      **64 B** |
|         **SingleNumber** | **Job-XQYQUJ** |           **.NET 6.0** | **WithoutCache** |   **692.3 ns** |  **13.65 ns** |  **22.42 ns** |   **702.0 ns** |      **65 B** | **0.0200** |      **64 B** |
|                      |            |                    |              |            |           |           |            |           |        |           |
| **SingleNumberAsParams** | **Job-XQYQUJ** |           **.NET 6.0** |         **Prev** | **1,006.0 ns** |  **11.08 ns** |  **10.36 ns** | **1,009.0 ns** |     **124 B** | **0.0496** |     **160 B** |
| **SingleNumberAsParams** | **Job-XQYQUJ** |           **.NET 6.0** |      **Current** |   **225.0 ns** |   **2.63 ns** |   **2.46 ns** |   **224.6 ns** |     **124 B** | **0.0508** |     **160 B** |
| **SingleNumberAsParams** | **Job-XQYQUJ** |           **.NET 6.0** | **WithoutCache** |   **716.9 ns** |  **14.31 ns** |  **23.92 ns** |   **709.0 ns** |     **124 B** | **0.0505** |     **160 B** |
|                      |            |                    |              |            |           |           |            |           |        |           |
|        **RoundtripInts** | **Job-JLANOM** | **.NET Framework 4.8** |         **Prev** | **5,378.5 ns** | **107.42 ns** | **160.79 ns** | **5,341.4 ns** |      **80 B** | **0.2518** |     **794 B** |
|        **RoundtripInts** | **Job-JLANOM** | **.NET Framework 4.8** |      **Current** | **1,715.7 ns** |  **11.39 ns** |  **10.66 ns** | **1,719.1 ns** |      **80 B** | **0.0629** |     **201 B** |
|        **RoundtripInts** | **Job-JLANOM** | **.NET Framework 4.8** | **WithoutCache** | **3,169.3 ns** |  **60.93 ns** |  **72.54 ns** | **3,169.4 ns** |      **80 B** | **0.0610** |     **201 B** |
|                      |            |                    |              |            |           |           |            |           |        |           |
|       **RoundtripLongs** | **Job-JLANOM** | **.NET Framework 4.8** |         **Prev** | **6,179.5 ns** |  **89.13 ns** |  **83.37 ns** | **6,145.9 ns** |      **80 B** | **0.2823** |     **907 B** |
|       **RoundtripLongs** | **Job-JLANOM** | **.NET Framework 4.8** |      **Current** | **2,390.5 ns** |  **12.05 ns** |  **11.27 ns** | **2,390.9 ns** |      **80 B** | **0.0420** |     **136 B** |
|       **RoundtripLongs** | **Job-JLANOM** | **.NET Framework 4.8** | **WithoutCache** | **3,761.7 ns** |  **53.58 ns** |  **50.12 ns** | **3,772.0 ns** |      **80 B** | **0.0420** |     **136 B** |
|                      |            |                    |              |            |           |           |            |           |        |           |
|         **RoundtripHex** | **Job-JLANOM** | **.NET Framework 4.8** |         **Prev** | **7,386.4 ns** | **159.98 ns** | **471.69 ns** | **7,421.2 ns** |      **80 B** | **0.5646** |   **1,781 B** |
|         **RoundtripHex** | **Job-JLANOM** | **.NET Framework 4.8** |      **Current** | **4,387.1 ns** |  **47.14 ns** |  **44.09 ns** | **4,400.2 ns** |      **80 B** | **0.3738** |   **1,188 B** |
|         **RoundtripHex** | **Job-JLANOM** | **.NET Framework 4.8** | **WithoutCache** | **4,839.6 ns** |  **94.05 ns** | **134.88 ns** | **4,827.6 ns** |      **80 B** | **0.3738** |   **1,188 B** |
|                      |            |                    |              |            |           |           |            |           |        |           |
|         **SingleNumber** | **Job-JLANOM** | **.NET Framework 4.8** |         **Prev** | **1,398.3 ns** |   **8.49 ns** |   **7.94 ns** | **1,396.8 ns** |      **83 B** | **0.0458** |     **144 B** |
|         **SingleNumber** | **Job-JLANOM** | **.NET Framework 4.8** |      **Current** |   **295.6 ns** |   **2.92 ns** |   **2.74 ns** |   **295.1 ns** |      **83 B** | **0.0200** |      **64 B** |
|         **SingleNumber** | **Job-JLANOM** | **.NET Framework 4.8** | **WithoutCache** |   **952.7 ns** |  **18.67 ns** |  **29.61 ns** |   **940.2 ns** |      **83 B** | **0.0200** |      **64 B** |
|                      |            |                    |              |            |           |           |            |           |        |           |
| **SingleNumberAsParams** | **Job-JLANOM** | **.NET Framework 4.8** |         **Prev** | **1,435.6 ns** |  **22.71 ns** |  **21.24 ns** | **1,427.9 ns** |     **142 B** | **0.0763** |     **241 B** |
| **SingleNumberAsParams** | **Job-JLANOM** | **.NET Framework 4.8** |      **Current** |   **383.4 ns** |   **2.41 ns** |   **2.26 ns** |   **383.1 ns** |     **142 B** | **0.0505** |     **160 B** |
| **SingleNumberAsParams** | **Job-JLANOM** | **.NET Framework 4.8** | **WithoutCache** | **1,005.1 ns** |  **19.84 ns** |  **29.08 ns** |   **995.2 ns** |     **142 B** | **0.0496** |     **160 B** |



P.S. I had the same pull request before, but my commits had incorrect author email :(
The original is pull request is closed.