@jetersen 2022-04-27
This seems like a better implementation than I did in #65