@timerplayer 2022-04-28
@jetersen,

Thank you for your comments!

> @timerplayer Haven't looked through the entire code but why would DecodeSingle require any allocation? Since a long is on the stack.
The solution doesn't allocate any memory for `DecodeSingle` & `DecondeSingleLong` cases.
But, I had found a bug that leads some memory allocation if use `MinHashLength` option.
You can check the change in the commit above.

I haven't fix the memory allocation in `Hex` methods.
Also, there is an issue in Decode & Encode method for int arrays.
It allocates twice more that need, because of `Int32` array convert procedure to `Int64`.