@ullmark 2022-05-17
@manigandham this PR has been here for a while. I think I've already mentioned my opinions re optimisations before... Of course the code shouldn't be slow, but I try to not jump into doing optimisations of code before there is an actual confirmed issue with the code being slow.

Refactoring code to make it easier or more maintainable is always good though. But my concern is that this library has gone from what I started with **1** class/files, to in this PR **40+** files without any new real features added. So my concern is: Is the optimisations here worth the added complexity. 🤷‍♂️   It's basically a rewrite, that has to be maintained by someone, and we don't have many maintainers.