@jetersen 2022-05-19
@ullmark there is also my PR #65 which tries to stay true to the project.

It has gotten rid of the biggest memory allocation consumer being the string.Split at the same time managing to do zero memory allocation for singledecode and encode.

It also comes with slight speed improve since it uses Span.