@manigandham 2022-05-22
v1.6.0 released https://www.nuget.org/packages/Hashids.net/1.6.0